Refactor update UI
